start state refactor
